backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
groq==0.11.0
pandas==2.2.2
httpx>=0.27.0
httpx>=0.24.0
